requirements.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
slowapi>=0.1.9
openai>=1.108,<2.0
pydantic>=2.11,<3.0
python-multipart>=0.0.20,<1.0
pytest>=8.4,<9.0
jieba>=0.42,<1.0
python-dotenv>=1.1,<2.0
imagehash>=4.3
Pillow>=10.0
rapidocr-onnxruntime>=1.3
